+++ subsetter/sampler.py
@@ -3,18 +3,15 @@ import json
 import logging
 import os
 import re
-from typing import Any, Callable, Dict, List, Literal, Optional, Set, Tuple, Union
+from typing import Any, Dict, List, Literal, Optional, Set, Tuple, Union
 
 import sqlalchemy as sa
 from pydantic import BaseModel, Field
+from sqlalchemy.ext.compiler import compiles
+from sqlalchemy.sql.expression import ClauseElement, Executable
 from typing_extensions import Annotated
 
-from subsetter.common import (
-    DEFAULT_DIALECT,
-    DatabaseConfig,
-    SQLDialectEncoder,
-    parse_table_name,
-)
+from subsetter.common import DatabaseConfig, parse_table_name
 from subsetter.filters import FilterConfig, FilterView, FilterViewChain
 from subsetter.metadata import DatabaseMetadata
 from subsetter.planner import SubsetPlan
@@ -30,6 +27,51 @@ except ImportError:
 LOGGER = logging.getLogger(__name__)
 SOURCE_BUFFER_SIZE = 1024
 DESTINATION_BUFFER_SIZE = 1024
+
+
+# pylint: disable=too-many-ancestors,abstract-method
+class TemporaryTable(Executable, ClauseElement):
+    inherit_cache = True
+
+    TEMP_ID = "_tmp_subsetter_"
+
+    def __init__(self, schema: str, name: str, select: sa.Select) -> None:
+        self.schema = schema
+        self.name = name
+        self.select = select
+        self.table_obj: sa.Table
+
+    def _temporary_table_compile_generic(self, compiler, **_) -> str:
+        name = self.name + self.TEMP_ID
+        schema_enc = compiler.dialect.identifier_preparer.quote(self.schema)
+        name_enc = compiler.dialect.identifier_preparer.quote(name)
+        select_stmt = compiler.process(self.select)
+        self.table_obj = sa.Table(
+            name,
+            sa.MetaData(),
+            *(sa.Column(col.name, col.type) for col in self.select.selected_columns),
+            schema=self.schema,
+        )
+        return f"CREATE TEMPORARY TABLE {schema_enc}.{name_enc} AS {select_stmt}"
+
+    def _temporary_table_compile_postgres(self, compiler, **_) -> str:
+        """
+        Postgres creates temporary tables in a special schema. We make the table
+        name incorporate the schema name to compensate and avoid collisions.
+        """
+        name = self.schema + self.TEMP_ID + self.name
+        name_enc = compiler.dialect.identifier_preparer.quote(name)
+        select_stmt = compiler.process(self.select)
+        self.table_obj = sa.Table(
+            name,
+            sa.MetaData(),
+            *(sa.Column(col.name, col.type) for col in self.select.selected_columns),
+        )
+        return f"CREATE TEMPORARY TABLE {name_enc} AS {select_stmt}"
+
+
+compiles(TemporaryTable, "postgresql")(TemporaryTable._temporary_table_compile_postgres)
+compiles(TemporaryTable)(TemporaryTable._temporary_table_compile_generic)
 
 
 def _multiply_column(
@@ -90,22 +132,29 @@ class SamplerOutput(abc.ABC):
     ) -> None:
         pass
 
-    def insert_order(self, tables: List[str], *, truncate: bool = False) -> List[str]:
-        # pylint: disable=unused-argument
-        return tables
+    def truncate(self) -> None:
+        """Delete any existing data that could interfere with output destination"""
+
+    @abc.abstractmethod
+    def insert_order(self) -> List[str]:
+        """Return the order to insert data that respects foreign key relationships"""
 
     @staticmethod
-    def from_config(config: Any) -> "SamplerOutput":
+    def from_config(config: Any, tables: List[str]) -> "SamplerOutput":
         if isinstance(config, DirectoryOutputConfig):
-            return DirectoryOutput(config)
+            return DirectoryOutput(config, tables)
         if isinstance(config, DatabaseOutputConfig):
-            return DatabaseOutput(config)
+            return DatabaseOutput(config, tables)
         raise RuntimeError("Unknown config type")
 
 
 class DirectoryOutput(SamplerOutput):
-    def __init__(self, config: DirectoryOutputConfig) -> None:
+    def __init__(self, config: DirectoryOutputConfig, tables: List[str]) -> None:
         self.directory = config.directory
+        self.tables = list(tables)
+
+    def insert_order(self) -> List[str]:
+        return self.tables
 
     def output_result_set(
         self,
@@ -139,13 +188,40 @@ class DirectoryOutput(SamplerOutput):
 
 
 class DatabaseOutput(SamplerOutput):
-    def __init__(self, config: DatabaseOutputConfig) -> None:
+    # Pylint getting weirdly confused about the self.meta member
+    # pylint: disable=no-member
+
+    def __init__(self, config: DatabaseOutputConfig, tables: List[str]) -> None:
         self.engine = config.database_engine(env_prefix="SUBSET_DESTINATION_")
-        self.sql_enc = SQLDialectEncoder.from_dialect(
-            config.dialect or os.getenv("SUBSET_DESTINATION_DIALECT") or DEFAULT_DIALECT  # type: ignore
-        )
         self.remap = config.remap
-        self.table_bind_processors: Dict[str, Dict[str, Callable[[Any], Any]]] = {}
+
+        self.table_remap: Dict[str, str] = {}
+        for table in tables:
+            remapped_table = ".".join(self._remap_table(*parse_table_name(table)))
+            if self.table_remap.setdefault(remapped_table, table) != table:
+                raise ValueError(
+                    f"Multiple tables remapped to the same name {remapped_table}"
+                )
+
+        self.meta, self.additional_tables = DatabaseMetadata.from_engine(
+            self.engine,
+            list(self.table_remap),
+            close_backward=True,
+        )
+
+    def truncate(self) -> None:
+        for schema, table_name in self.additional_tables:
+            LOGGER.info(
+                "Found additional table %s.%s to truncate",
+                schema,
+                table_name,
+            )
+
+        with self.engine.connect() as conn:
+            for table in reversed(self.meta.toposort()):
+                LOGGER.info("Truncating table %s", table)
+                conn.execute(sa.delete(table.table_obj))
+                conn.commit()
 
     def _remap_table(self, schema: str, table_name: str) -> Tuple[str, str]:
         table = f"{schema}.{table_name}"
@@ -153,60 +229,18 @@ class DatabaseOutput(SamplerOutput):
             table = re.sub(remap_pattern.search, remap_pattern.replace, table)
         return parse_table_name(table)
 
-    def insert_order(self, tables: List[str], *, truncate: bool = True) -> List[str]:
-        table_remap: Dict[str, str] = {}
-        for table in tables:
-            remapped_table = ".".join(self._remap_table(*parse_table_name(table)))
-            if table_remap.setdefault(remapped_table, table) != table:
-                raise ValueError(
-                    f"Multiple tables remapped to the same name {remapped_table}"
-                )
-
-        meta, additional_tables = DatabaseMetadata.from_engine(
-            self.engine,
-            list(table_remap),
-            close_backward=truncate,
-        )
-        for table in table_remap:
-            table_meta = meta.tables.get(parse_table_name(table))
-            if table_meta is None:
+    def insert_order(self) -> List[str]:
+        result = []
+        insert_order_mapped = [str(table) for table in self.meta.toposort()]
+        for table in insert_order_mapped:
+            remapped_table = self.table_remap.get(table)
+            if remapped_table:
+                result.append(remapped_table)
+            else:
                 LOGGER.warning(
                     "Database does not have table %s, will not sample", table
                 )
-            else:
-                bind_processors = {}
-                for col_name, col_data in table_meta.columns.items():
-                    if bind_processor := col_data.type_.bind_processor(
-                        self.engine.dialect
-                    ):
-                        bind_processors[col_name] = bind_processor
-                self.table_bind_processors[table] = bind_processors
-        for schema, table_name in additional_tables:
-            LOGGER.info(
-                "Found additional unsampled table %s.%s to truncate",
-                schema,
-                table_name,
-            )
-
-        insert_order_mapped = [str(table) for table in meta.toposort()]
-
-        if truncate:
-            with self.engine.connect() as conn:
-                for table in reversed(insert_order_mapped):
-                    schema, table_name = parse_table_name(table)
-                    LOGGER.info("Truncating table %s.%s", schema, table_name)
-                    conn.execute(
-                        sa.text(
-                            f"DELETE FROM {self.sql_enc.table_name(schema, table_name)} WHERE 1=1"
-                        )
-                    )
-                    conn.commit()
-
-        return [
-            remapped_table
-            for table in insert_order_mapped
-            if (remapped_table := table_remap.get(table, ""))
-        ]
+        return result
 
     def output_result_set(
         self,
@@ -220,32 +254,21 @@ class DatabaseOutput(SamplerOutput):
         column_multipliers: Optional[Set[str]] = None,
     ) -> None:
         columns_out = filter_view.columns_out if filter_view else columns
-        column_index = {col_name: index for index, col_name in enumerate(columns_out)}
         multiplied_indexes = [
             ind
             for ind, col_name in enumerate(columns_out)
             if col_name in (column_multipliers or set())
         ]
         schema, table_name = self._remap_table(schema, table_name)
-        insert_query = (
-            f"INSERT INTO {self.sql_enc.table_name(schema, table_name)} "
-            f"({self.sql_enc.column_list(columns_out)}) VALUES "
-        )
         buffer: List[tuple] = []
-        bind_processors = [
-            (ind, bind_processor)
-            for col_name, bind_processor in self.table_bind_processors.get(
-                f"{schema}.{table_name}", {}
-            ).items()
-            if (ind := column_index.get(col_name)) is not None
-        ]
+
+        table = self.meta.tables[(schema, table_name)]
 
         def _flush_buffer():
-            query = f"{insert_query}{','.join(f':{ind}' for ind in range(len(buffer)))}"
             with self.engine.connect() as conn:
                 conn.execute(
-                    sa.text(query),
-                    {str(ind): row for ind, row in enumerate(buffer)},
+                    sa.insert(table.table_obj),
+                    [dict(zip(columns_out, row)) for row in buffer],
                 )
                 conn.commit()
             buffer.clear()
@@ -257,8 +280,6 @@ class DatabaseOutput(SamplerOutput):
                     out_row[index] = _multiply_column(
                         out_row[index], multiplier, iteration
                     )
-                for index, bind_processor in bind_processors:
-                    out_row[index] = bind_processor(out_row[index])
                 buffer.append(tuple(out_row))
                 if len(buffer) > DESTINATION_BUFFER_SIZE:
                     _flush_buffer()
@@ -272,12 +293,6 @@ class Sampler:
         self.source_engine = self.config.source.database_engine(
             env_prefix="SUBSET_SOURCE_"
         )
-        self.sql_enc = SQLDialectEncoder.from_dialect(
-            self.config.source.dialect  # type: ignore
-            or os.getenv("SUBSET_SOURCE_DIALECT")
-            or DEFAULT_DIALECT
-        )
-        self.output = SamplerOutput.from_config(config.output)
 
     def sample(self, plan: SubsetPlan, *, truncate: bool = False) -> None:
         meta, _ = DatabaseMetadata.from_engine(self.source_engine, list(plan.queries))
@@ -285,33 +300,33 @@ class Sampler:
             meta.infer_missing_foreign_keys()
         self._validate_filters(meta)
 
-        insert_order = self.output.insert_order(list(plan.queries), truncate=truncate)
         table_column_multipliers = self._get_multiplied_columns(
             meta, list(plan.queries)
         )
 
-        with self.source_engine.execution_options().connect() as conn:
-            self._materialize_tables(conn, plan)
-            self._copy_results(conn, meta, plan, insert_order, table_column_multipliers)
+        output = SamplerOutput.from_config(self.config.output, list(plan.queries))
+        insert_order = output.insert_order()
+        if truncate:
+            output.truncate()
 
-    def _materialize_tables(self, conn, plan: SubsetPlan) -> None:
+        with self.source_engine.execution_options().connect() as conn:
+            self._materialize_tables(meta, conn, plan)
+            self._copy_results(
+                output, conn, meta, plan, insert_order, table_column_multipliers
+            )
+
+    def _materialize_tables(
+        self, meta: DatabaseMetadata, conn: sa.Connection, plan: SubsetPlan
+    ) -> None:
         for table, query in plan.queries.items():
             if not query.materialize:
                 continue
             schema, table_name = parse_table_name(table)
-
             LOGGER.info("Materializing sample for %s", table)
 
-            params: Dict[str, Any] = {}
-            sample_sql = query.build(self.sql_enc, params)
-
+            ttbl = TemporaryTable(schema, table_name, query.build(meta))
             try:
-                result = conn.execute(
-                    sa.text(
-                        self.sql_enc.make_temp_table(schema, table_name, sample_sql)
-                    ),
-                    params,
-                )
+                result = conn.execute(ttbl)
             except Exception as exc:  # pylint: disable=broad-exception-caught
                 # Some client/server combinations report a read-only error even though the temporary
                 # table creation actually succeeded. We'll just swallow the error here and if there
@@ -319,6 +334,7 @@ class Sampler:
                 if "--read-only" not in str(exc):
                     raise
             else:
+                meta.temp_tables[(schema, table_name)] = ttbl.table_obj
                 LOGGER.info(
                     "Materialized %d rows for %s.%s in temporary table",
                     result.rowcount,
@@ -328,6 +344,7 @@ class Sampler:
 
     def _copy_results(
         self,
+        output: SamplerOutput,
         conn,
         meta: DatabaseMetadata,
         plan: SubsetPlan,
@@ -343,17 +360,16 @@ class Sampler:
 
             LOGGER.info("Sampling %s.%s ...", schema, table_name)
 
-            params: Dict[str, Any] = {}
             if query.materialize:
-                query_sql = f"SELECT * FROM {self.sql_enc.table_name(schema, table_name, sampled=True)}"
+                query_stmt = sa.select(meta.temp_tables[(schema, table_name)])
             else:
-                query_sql = query.build(self.sql_enc, params)
+                query_stmt = query.build(meta)
 
-            LOGGER.info("Sampling with query %r", query_sql)
+            LOGGER.info("Sampling with query %r", query_stmt)
             result = conn.execution_options(
                 stream_results=True,
                 yield_per=SOURCE_BUFFER_SIZE,
-            ).execute(sa.text(query_sql), params)
+            ).execute(query_stmt)
             columns = result.keys()
 
             filter_view = FilterViewChain.construct_filter(
@@ -370,18 +386,12 @@ class Sampler:
                     rows += 1
                     yield row
 
-            table_columns = meta.tables[(schema, table_name)].columns
-            processed_results = self.sql_enc.result_processor(
-                [table_columns[column].type_ for column in columns],
-                result,
-            )
-
             column_multipliers = table_column_multipliers.get(table)
-            self.output.output_result_set(
+            output.output_result_set(
                 schema,
                 table_name,
                 columns,
-                _count_rows(processed_results),
+                _count_rows(result),
                 filter_view=filter_view,
                 multiplier=(
                     1
@@ -403,7 +413,10 @@ class Sampler:
                 LOGGER.warning("Found filters for unknown table %s", table)
                 continue
 
-            FilterViewChain.construct_filter(tbl.columns, filters)
+            FilterViewChain.construct_filter(
+                tuple(column.name for column in tbl.table_obj.columns),
+                filters,
+            )
 
     def _get_multiplied_columns(
         self, meta: DatabaseMetadata, tables: List[str]
@@ -470,11 +483,12 @@ class Sampler:
 
         for table_name in tables:
             table = meta.tables[parse_table_name(table_name)]
+            col_map = {column.name: column for column in table.table_obj.columns}
 
             # Verify multiplied columns are integer
             for col_name in result.get(table_name, set()):
-                col = table.columns[col_name]
-                if not issubclass(col.type_.python_type, int):  # type: ignore
+                col = col_map[col_name]
+                if not issubclass(col.type.python_type, int):  # type: ignore
                     raise ValueError(
                         f"Primary key column {table_name}.{col_name} "
                         "must be integral when using multiplicity"